Move Avalon python worker crypto files to crypto directory.
- Move crypto files and update the module import path.
